Best model for each criterion
Added option to save the best model for each criterion after epoch loop.
This done buy tracking the best epoch for the each criterion.
At the end copy epoch model to the model with criterion name added.
If training is interrupted best epoch per criterion log is saved in
checkpoint and reload on training resume.